refactor(tables): migrate row parse/validation to the column-type registry
coerceValueToColumnType and validateRowAgainstSchema's switches are
replaced by delegates to the registry's parse/isValidValue, so a new
column type's coercion and shape-check rules only need to be declared
once. The now-dead optionIds helper is removed, and
resolveSelectOptionId/splitMultiSelectInput move to select-values.ts
(their tests move with them) now that the registry depends on them
living there instead.
